third_party/blink/web_tests/media/mediasession/mojo/file-image-removed.html

<!DOCTYPE html>
<title>MediaSession Mojo Test</title>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="file:///gen/layout_test_data/mojo/public/js/mojo_bindings.js"></script>
<script src="file:///gen/third_party/blink/public/mojom/mediasession/media_session.mojom.js"></script>
<script src="resources/mediasessionservice-mock.js"></script>
<script src="resources/utils.js"></script>
<script>

async_test(function(t) {
  let m = mediaSessionServiceMock;
  var metadata = new MediaMetadata({
    artwork: [
      { src: "file:///foo/bar.jpg", type: "image/jpeg"}
    ]});
  var expectedMetadata = new MediaMetadata({});

  m.setMetadataCallback(t.step_func(function(receivedMetadata) {
    assert_metadata_equals(expectedMetadata, receivedMetadata);
    t.done();
  }));
  window.navigator.mediaSession.metadata = metadata;
}, "test that null MediaMetadata is correctly propagated");

</script>
